+++ R/gam_tidiers.R
@@ -7,13 +7,18 @@
 #' 
 #' The "augment" method is handled by \link{lm_tidiers}.
 #'
-#' @param x gam object
+#' @param x gam or Gam object
+#' @param parametric logical. Return parametric coefficients (\code{TRUE}) or 
+#' information about smooth terms (\code{FALSE})?
 #' 
 #' @template boilerplate
 #' 
-#' @return \code{tidy.gam} returns the tidied output of the
-#' parametric ANOVA for gam package  with one row for each term in the formula.
-#' The columns match those in \link{anova_tidiers}.
+#' @return \code{tidy.gam} called on an object from the gam package,
+#' or an object from the mgcv package with \code{parametric = FALSE}, returns the 
+#' tidied output of the parametric ANOVA with one row for each term in the formula.
+#' The columns match those in \link{anova_tidiers}. 
+#' \code{tidy.gam} called on a gam object from the mgcv package with 
+#' \code{parametric = TRUE} returns the fixed coefficients.
 #'
 #' @name gam_tidiers
 #' 
@@ -30,18 +35,17 @@
 #' }
 #'
 #' @export
-tidy.gam <- function(x, ...) {
-    
+tidy.gam <- function(x, parametric = FALSE, ...) {
     if(is_mgcv(x)){
-        tidy_mcgv(x)
+        tidy_mcgv(x, parametric)
     }else{
-        tidy_gam(x)
+        tidy.Gam(x)
     }
-
 }
 
 is_mgcv <- function(x) {
-    #figure out if gam is from package gam or mcgv
+    # figure out if gam is from package gam or mcgv
+    # As of gam 1.15 this is no longer necessary since gam's class is now named Gam
     mgcv_names <- c("linear.predictors", "converged", "sig2", "edf", "edf1", 
                     "hat", "boundary", "sp", "nsdf", "Ve", "Vp", "rV", 
                     "gcv.ubre", "scale.estimated", "var.summary", 
@@ -50,16 +54,24 @@ is_mgcv <- function(x) {
     
 }
 
-tidy_gam <- function(x) {
+#' @rdname gam_tidiers
+#' @export
+tidy.Gam <- function(x, ...) {
     # return the output of the parametric ANOVA
     tidy(summary(x)$parametric.anova) 
 }
 
-tidy_mcgv <- function(x) {
-    sx <- summary(x)$s.table
-    sx <- as.data.frame(sx)
-    class(sx) <- c("anova", "data.frame")
-    tidy(sx)
+tidy_mcgv <- function(x, parametric = FALSE) {
+    if (parametric) {
+        px <- summary(x)$p.table
+        px <- as.data.frame(px) 
+        fix_data_frame(px, c("estimate", "std.error", "statistic", "p.value"))  
+    } else {
+        sx <- summary(x)$s.table
+        sx <- as.data.frame(sx)
+        class(sx) <- c("anova", "data.frame")
+        tidy(sx)
+    }
 }
 
 #' @rdname gam_tidiers
@@ -79,12 +91,14 @@ glance.gam <- function(x, ...) {
     if(is_mgcv(x)){
         glance_mcgv(x)
     }else{
-        glance_gam(x)
+        glance.Gam(x)
     }
 }
 
 
-glance_gam <- function(x) {
+#' @rdname gam_tidiers
+#' @export 
+glance.Gam <- function(x, ...) {
     s <- summary(x)
     ret <- data.frame(df = s$df[1])
     finish_glance(ret, x)